Add files via upload
This will not work anywhere outside of my university's server; it depends on their simulated environment. This program gets information from a function called _getPacket; finds the index of the name provided by the function via getDest, and again if the second name given is also a student; then applies the logic outlined in out assignment.