Add files via upload
HW6main.s is the main file provided by my instructor. StudentCode.s is a function called by the previous file whose purpose was to alter the main file via stack smashing. I accomplished this by loading in the value of a NOP - as specified by the fictional assembly language manual for the class - and placing it in the location of the function I wished to ignore. I then placed the start of my string in the location the program expected the start of my instructors string.